refactor: chat 도메인 리팩토링
- 책임분리 수행
